Revert "chore: Failling trial for Android/iOS. Keeping the commit just for future reference"
This reverts commit 02eada538fc9dd549951b68fa2d77ec35918e080.
