+++ src/Uno.UI/UI/Xaml/LeaveParams.cs
@@ -1,4 +1,5 @@
-﻿namespace Uno.UI.Xaml;
+﻿#if UNO_HAS_ENHANCED_LIFECYCLE
+namespace Uno.UI.Xaml;
 
 internal struct LeaveParams
 {
@@ -14,3 +15,4 @@ internal struct LeaveParams
 		IsLive = isLive;
 	}
 }
+#endif
