+++ src/Uno.UI/UI/Xaml/UIElement.Android.cs
@@ -18,7 +18,6 @@ using Point = Windows.Foundation.Point;
 using Rect = Windows.Foundation.Rect;
 using Java.Interop;
 using Microsoft.UI.Xaml.Markup;
-using Uno.UI.Xaml;
 
 namespace Microsoft.UI.Xaml
 {
@@ -66,7 +65,6 @@ namespace Microsoft.UI.Xaml
 			if (view is UIElement uiElement)
 			{
 				OnChildManagedViewAddedOrRemoved(uiElement);
-				uiElement.Enter(FindNameScope(uiElement), default, default);
 			}
 			else
 			{
@@ -82,7 +80,6 @@ namespace Microsoft.UI.Xaml
 			if (view is UIElement uiElement)
 			{
 				OnChildManagedViewAddedOrRemoved(uiElement);
-				uiElement.Leave(FindNameScope(uiElement), default);
 			}
 			else
 			{
