+++ src/Uno.UI/UI/Xaml/UIElement.cs
@@ -1434,25 +1434,6 @@ namespace Microsoft.UI.Xaml
 			ProtectedCursor = cursor;
 		}
 
-		internal static INameScope FindNameScope(DependencyObject dependencyObject)
-		{
-			// In many cases, this will not traverse the entire visual tree, as the NameScope is often
-			// present on each element as the DP is inherited. However, it can happen that the DP inheritance
-			// didn't yet happen, in which case we'll traverse the tree up.
-			var current = dependencyObject;
-			while (current is not null)
-			{
-				if (NameScope.GetNameScope(current) is { } nameScope)
-				{
-					return nameScope;
-				}
-
-				current = VisualTreeHelper.GetParent(current);
-			}
-
-			return null;
-		}
-
 		/// <summary>
 		/// This event is not yet implemented in Uno Platform.
 		/// </summary>
