+++ src/Uno.UI/UI/Xaml/UIElement.mux.cs
@@ -616,27 +616,18 @@ namespace Microsoft.UI.Xaml
 			return new Rect();
 		}
 
+#if UNO_HAS_ENHANCED_LIFECYCLE
 		// Doesn't exactly match WinUI code.
 		internal virtual void Enter(INameScope? nameScope, EnterParams @params, int depth)
 		{
-			// Enter does nothing on platforms with non-enhanced lifecycle.
-			// Its only purpose is to register names via the FrameworkElement override.
-#if UNO_HAS_ENHANCED_LIFECYCLE
 			Depth = depth;
 
 			if (@params.IsLive)
 			{
 				IsActiveInVisualTree = true;
 			}
-#endif
 
-#if UNO_HAS_ENHANCED_LIFECYCLE
-			var children = _children;
-#else
-			var children = this.GetChildren();
-#endif
-
-			foreach (var child in children)
+			foreach (var child in _children)
 			{
 				if (child == this)
 				{
@@ -648,14 +639,9 @@ namespace Microsoft.UI.Xaml
 					continue;
 				}
 
-#if UNO_HAS_ENHANCED_LIFECYCLE
 				child.Enter(nameScope, @params, depth + 1);
-#else
-				(child as UIElement)?.Enter(nameScope, @params, depth + 1);
-#endif
 			}
 
-#if UNO_HAS_ENHANCED_LIFECYCLE
 			if (@params.IsLive)
 			{
 				var core = this.GetContext();
@@ -664,19 +650,11 @@ namespace Microsoft.UI.Xaml
 
 			// Make sure that we propagate OnDirtyPath bits to the new parent.
 			SetLayoutFlags(LayoutFlag.MeasureDirty | LayoutFlag.ArrangeDirty);
-#endif
 		}
 
 		internal virtual void Leave(INameScope? nameScope, LeaveParams @params)
 		{
-			var children =
-#if UNO_HAS_ENHANCED_LIFECYCLE
-				_children;
-#else
-				this.GetChildren();
-#endif
-
-			foreach (var child in children)
+			foreach (var child in _children)
 			{
 				if (child == this)
 				{
@@ -688,14 +666,9 @@ namespace Microsoft.UI.Xaml
 					continue;
 				}
 
-#if UNO_HAS_ENHANCED_LIFECYCLE
 				child.Leave(nameScope, @params);
-#else
-				(child as UIElement)?.Leave(nameScope, @params);
-#endif
 			}
 
-#if UNO_HAS_ENHANCED_LIFECYCLE
 			if (IsActiveInVisualTree)
 			{
 				if (IsLoaded)
@@ -716,7 +689,8 @@ namespace Microsoft.UI.Xaml
 
 				var eventManager = this.GetContext().EventManager;
 				eventManager.RemoveRequest(this);
-#endif
+			}
 		}
+#endif
 	}
 }
